+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Replaced "semi-supervised" with "prediction-powered" everywhere in the docs and code
 - Rewrote CLT and Bootstrap confidence intervals to cache lower/upper bounds for a given confidence level
 
+### 🐛 Fixed
+- Handling of NaN values in `StratifiedClassicalMeanEstimator`
+
 ### 💛 Contributors
 
 ---